requirements.txt
# LLM Providers
openai>=1.50.0
google-generativeai>=0.8.0
google-genai>=1.0.0
anthropic>=0.40.0
google-cloud-aiplatform>=1.70.0
vertexai>=1.70.0

# Core dependencies
Pillow>=10.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
PyYAML>=6.0.0
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.0

# Schema validation
openapi-schema-validator>=0.6.0
referencing>=0.31.0

# USDM validation (official CDISC package)
usdm>=0.60.0

# Testing
pytest>=8.0.0
